feat: professor can now manage his office hours for students to book slots

+++ backend/src/main/java/com/university/backend/controller/StaffDirectoryController.java
@@ -42,7 +42,7 @@ public class StaffDirectoryController {
 
     @PostMapping("/slots/{slotId}/book")
     public ResponseEntity<?> bookOfficeHourSlot(
-            @PathVariable Long slotId,
+            @PathVariable Integer slotId,
             @RequestBody BookSlotRequest request,
             HttpSession session) {
         try {
